Fix (another) integer type mismatch on 64-bit systems
Similar to previous commit, use int type as the first parameter to
printf field width specifier (%.*s).

+++ tig.c
@@ -1394,7 +1394,7 @@ update_view_title(struct view *view)
 
 		if (minsize < view->width)
 			refsize = view->width - minsize + 7;
-		string_format_from(buf, &bufpos, " %.*s", refsize, view->ref);
+		string_format_from(buf, &bufpos, " %.*s", (int) refsize, view->ref);
 	}
 
 	if (statelen && bufpos < view->width) {
